test(fixtures): move bench scene generation into dsp_fixtures
The null test, the benches and the future block-time gate must measure
the same scenes; duplicated generators drift. Pure move plus a new
render_blocks helper for multi-block capture.

Bench re-baseline (render_steady, x86_64 Linux, criterion median):

  objects  before      after       change
        1  ~4.457 us   4.394 us    -1.40%
        8   7.520 us   7.528 us    +0.16%  (no change detected)
       16  11.348 us  11.175 us    -1.18%
       32  18.295 us  18.245 us    -0.09%  (no change detected)
       64  33.106 us  32.919 us    +0.13%  (no change detected)
      118  57.037 us  56.139 us    -1.50%

Every figure moved by well under 5%, confirming this was a pure move.
The "before" column was measured from a worktree at the parent commit
sharing the same criterion data directory.
